% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_for_faulty_xml}
\alias{check_for_faulty_xml}
\title{Detect Faulty XML Request}
\usage{
check_for_faulty_xml(request_obj)
}
\value{
Message to the console.
}
\description{
Searches the results content for the tag "<FaultString>". If it
is found it gives the user a message and exits the function.
}
\keyword{internal}
