% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_for_invalid_xml}
\alias{check_for_invalid_xml}
\title{Detect Invalid XML Request}
\usage{
check_for_invalid_xml(request_obj)
}
\value{
Message to the console.
}
\description{
Searches the results content for the phrase "Invalid XML Request". If it
is found it gives the user a message and exits the function.
}
\keyword{internal}
