% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_request_status}
\alias{check_request_status}
\title{Check Request Status}
\usage{
check_request_status(request_obj)
}
\value{
If status code is not 200, a message to console
}
\description{
This function is called by other functions that submit jobs to
the Acoustic/Silverpop API. It checks the status code
returned and tells the user if there was an error code and
exits the function.
}
\keyword{internal}
