% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_job_id}
\alias{get_job_id}
\title{Get Job Id of Submitted Job}
\usage{
get_job_id(request_obj)
}
\arguments{
\item{request_obj}{Name of the object returned from API call,
should always be "request".}
}
\value{
A vector with the Job Id.
}
\description{
This function is called by other functions that submit jobs to
the Acoustic/Silverpop API. It extracts the Job Id from the XML
returned by the API call.
}
\details{
Job results are available as exports in the Silverpop portal by
going to Resources -> Data Jobs.

It is not recommended that these authentication parameters be
stored directly in your code. There are various methods and
packages available that are more secure; this package does not
require you to use any one in particular.
}
\keyword{internal}
